# docs/openapi.yaml
openapi: 3.0.3
info:
  title: weather_api
  description: |
    SBS weather_api HTTP 계약 (실구현 기준).

    - Producer 카탈로그: `skills/weather-api-catalog`
    - HGT500 상세 클라이언트 플로우: `docs/kim_hgt500_frontend_api_spec.md`
    - HGT500 packed PNG 변환/복호화: `skills/kim-hgt500-png-pipeline`
    - 태풍/admin API는 이 서버에 없음 (별도 typhoon API 참고)

    미구현(PRD only, 문서화하지 않음): `POST /api/hgt500/datasets`, `GET /api/hgt500/jobs/{jobId}`
  version: 1.0.0
  contact:
    name: weather_api
servers:
  - url: http://localhost:3010
    description: Local (listen 0.0.0.0:3010)
  - url: https://weather-map.sbs.co.kr
    description: Production
...
paths:
  /api/hgt500/latest:
    get:
      tags: [hgt500]
      summary: 최신 KIM HGT500 dataset pointer
      description: |
        최신 성공 dataset의 mutable pointer.
        animation은 이 응답만으로 만들지 말고 `manifestUrl`의 frames를 사용한다.
      operationId: getHgt500Latest
      responses:
        '200':
          description: 최신 dataset pointer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hgt500DatasetPointer'
              example:
                datasetId: kim-glob-hgt500-2026070100
                tmfc: '2026070100'
                status: succeeded
                sourceFormat: kim-api-text
                downsampleFactor: 3
                analysisTime: '2026-07-01T00:00:00Z'
                validTimeStart: '2026-07-01T00:00:00Z'
                validTimeEnd: '2026-07-04T00:00:00Z'
                sourceForecastIntervalMinutes: 180
                outputFrameIntervalMinutes: 10
                frameCount: 433
                manifestUrl: /datasets/kim-glob-hgt500-2026070100/manifest.json
        '404':
          description: 생성된 dataset 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: No KIM HGT500 dataset is available
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWithDetails'
  /api/hgt500/datasets:
    get:
      tags: [hgt500]
      summary: KIM HGT500 dataset 목록
      description: |
        디스크의 성공 dataset을 스캔한다. `Cache-Control: no-store`.
        정렬은 `tmfc` 내림차순. 항목 shape는 `/api/hgt500/latest`와 동일.
      operationId: listHgt500Datasets
      parameters:
        - name: tmfc
          in: query
          schema:
            type: string
            pattern: '^\d{10}$'
          description: YYYYMMDDHH UTC. 해당 분석시각만
          example: '2026072800'
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: ISO-8601. valid window와 구간 겹침(inclusive)
          example: '2026-07-28T00:00:00Z'
        - name: to
          in: query
          schema:
            type: string
            format: date-time
          description: ISO-8601. valid window와 구간 겹침(inclusive)
          example: '2026-08-02T00:00:00Z'
        - name: intervalMinutes
          in: query
          schema:
            type: integer
          description: outputFrameIntervalMinutes 정확 일치
        - name: downsampleFactor
          in: query
          schema:
            type: integer
          description: downsampleFactor 정확 일치
        - name: sourceFormat
          in: query
          schema:
            type: string
          description: 예 kim-api-text
        - name: status
          in: query
          schema:
            type: string
            default: succeeded
          description: 기본 succeeded. `all`이면 status 필터 없음
      responses:
        '200':
          description: 목록 (empty items 가능)
          headers:
            Cache-Control:
              schema:
                type: string
                example: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hgt500DatasetList'
        '400':
          description: query 형식 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWithDetails'
  /api/hgt500/datasets/{datasetId}/manifest:
    get:
      tags: [hgt500]
      summary: manifest로 redirect
      description: |
        `302 Location: /datasets/{datasetId}/manifest.json`
        클라이언트는 보통 latest/list의 `manifestUrl`을 직접 fetch하는 편이 단순하다.
      operationId: redirectHgt500Manifest
      parameters:
        - $ref: '#/components/parameters/DatasetId'
      responses:
        '302':
          description: static manifest URL로 redirect
          headers:
            Location:
              schema:
                type: string
                example: /datasets/kim-glob-hgt500-2026070100/manifest.json
        '400':
          description: datasetId 형식 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Invalid datasetId
  /datasets/{datasetId}/manifest.json:
    get:
      tags: [static, hgt500]
      summary: Dataset manifest (static)
      description: |
        `@fastify/static` prefix `/datasets/`.
        Frame asset 경로는 manifest 기준 상대 경로.
        상세 schema는 `docs/kim_hgt500_frontend_api_spec.md` 참고.
      operationId: getHgt500ManifestStatic
      parameters:
        - $ref: '#/components/parameters/DatasetId'
      responses:
        '200':
          description: Manifest JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hgt500Manifest'
        '404':
          description: 파일 없음
...
  /api/aws/stations:
    get:
      tags: [aws]
      summary: AWS 지점 코드표 (stn_inf)
      description: |
        `kma_fetch/config/aws_stn_code_YYYYMMDD.json` (최신 날짜 suffix)을 반환한다.
        `LAW_ADDR_SIDO` / `LAW_ADDR_GUGUN` 및 최신 `STN_NAME`·좌표 lookup용.
        프레임 JSON(`/api/aws/min`)에는 LAW 필드를 저장하지 않으며, min 응답은 서버가 enrich한다.
      operationId: getAwsStations
      responses:
        '200':
          description: station catalog
          headers:
            Cache-Control:
              schema:
                type: string
                example: 'public, max-age=3600'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AwsStationsResponse'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWithDetails'
  /api/aws/min/pack:
    get:
      tags: [aws]
      summary: AWS 1분 TA packed timeline manifest
      description: |
        KST 1분 exact 범위의 TA Int16 LE pack manifest.
        2분 snap 없음. 최대 1440 frame. variable은 현재 `TA`만.
        Binary는 `data.url` (예: `/datasets/aws/ta/1m/{day}/ta.i16le`).
        layout: FRAME_MAJOR_STATION_MINOR, scale 0.1℃, missing -32768.
        과거 완결일(`complete:true`)은 immutable cache, today는 no-store.
      operationId: getAwsMinPack
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{12}$'
          example: '202608120000'
        - name: to
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{12}$'
          example: '202608122359'
        - name: variable
          in: query
          required: false
          schema:
            type: string
            enum: [TA]
            default: TA
        - name: force
          in: query
          required: false
          schema:
            type: string
            enum: ['0', '1', 'true', 'false']
          description: cache 무시하고 재생성
      responses:
        '200':
          description: pack manifest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AwsMinPackManifest'
        '400':
          description: 범위/variable 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 요청 범위에 원자료 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWithDetails'
  /api/aws/min/exact:
    get:
      tags: [aws]
      summary: AWS_MIN 1분 exact 단건 (debug/parity)
      description: |
        2분 snap 없이 `timestamp_kor` 파일만 읽는다. pack parity 확인용.
      operationId: getAwsMinExact
      parameters:
        - name: timestamp_kor
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{12}$'
          example: '202608121533'
      responses:
        '200':
          description: station 배열 (enrich 포함)
          content:
            application/json:
              schema:
                type: object
                properties:
                  timestamp_kor:
                    type: string
                  intervalMinutes:
                    type: integer
                    example: 1
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AwsStationRow'
        '400':
          description: 형식 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 파일 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/aws/min:
    get:
      tags: [aws]
      summary: AWS_MIN station JSON 단건
      description: |
        `in_data/aws/{yyyy-MM-dd}/AWS_MIN_{YYYYMMDDHHMM}.json`을 읽는다.
        기본: `timestamp_kor` 2분 nearest snap (호환).
        `intervalMinutes=1`이면 snap 없이 exact ( `/api/aws/min/exact` 와 동일).
        응답 `data[]`는 코드표 enrich.
      operationId: getAwsMin
      parameters:
        - name: timestamp_kor
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{12}$'
          description: YYYYMMDDHHMM KST
          example: '202604050000'
        - name: intervalMinutes
          in: query
          required: false
          schema:
            type: integer
            enum: [1, 2]
          description: 생략 또는 2 = 2분 snap, 1 = exact
      responses:
        '200':
          description: station 배열
          headers:
            Cache-Control:
              schema:
                type: string
                example: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AwsMinResponse'
        '400':
          description: timestamp_kor 누락/형식 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 해당 시각 파일 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  timestamp_kor:
                    type: string
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWithDetails'
...
  /ir105/{area}/{step}:
    get:
      tags: [ir105-json]
      summary: IR105 JSON 단건 (PostgreSQL)
      operationId: getIr105Json
      parameters:
        - $ref: '#/components/parameters/Ir105Area'
        - $ref: '#/components/parameters/Ir105Step'
        - name: timestamp_kor
          in: query
          required: true
          schema:
            type: string
          description: DB `observation_time_kor`와 일치하는 값
          example: '2025-03-01T00:00:00Z'
      responses:
        '200':
          description: ir105_json row
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: timestamp_kor 누락
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 데이터 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWithDetails'
...
components:
  parameters:
...
    Ir105Area:
      name: area
      in: path
      required: true
      schema:
        type: string
      example: ea
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ErrorWithDetails:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required: [error]
    Hgt500DatasetPointer:
      type: object
      properties:
        datasetId:
          type: string
          example: kim-glob-hgt500-2026070100
        tmfc:
          type: string
          description: YYYYMMDDHH UTC
        status:
          type: string
          example: succeeded
        sourceFormat:
          type: string
          example: kim-api-text
        downsampleFactor:
          type: integer
        analysisTime:
          type: string
          format: date-time
        validTimeStart:
          type: string
          format: date-time
        validTimeEnd:
          type: string
          format: date-time
        sourceForecastIntervalMinutes:
          type: integer
        outputFrameIntervalMinutes:
          type: integer
        frameCount:
          type: integer
        manifestUrl:
          type: string
          description: 상대 또는 절대 path. 보통 `/datasets/{id}/manifest.json`
...